Create CountingBits.java
Question Link: https://leetcode.com/problems/counting-bits/submissions/
Approach: First approach is simply using the solution of the number which is it's half. Ex: for 9, which is 1001, and it's half is 4 which is 100. So, just an extra bit is added in 9, which is found by 9%2, and the remaining bits are same.
Approach 2: Found a pattern and coded on top of that. Pattern:
Number of 1's given sequentially: 0 1 1 2 1 2 2 3 1 2 2 3 2 3 3 4 1 2 2 3  2 3 3 4  2 3 3 4  3 4 4 5 1 2 2 3  2 3 3 4  2 3 3 4  3 4 4 5  2 3 3 4  3 4 4 5  3 4 4 5  4 5 5 6
This can be viewed as:
0
1
1 2
1 2 2 3
1 2 2 3  2 3 3 4
1 2 2 3  2 3 3 4  2 3 3 4  3 4 4 5
1 2 2 3  2 3 3 4  2 3 3 4  3 4 4 5  2 3 3 4  3 4 4 5  3 4 4 5  4 5 5 6
1 2 2 3  2 3 3 4  2 3 3 4  3 4 4 5  2 3 3 4  3 4 4 5  3 4 4 5  4 5 5 6  2 3 3 4  3 4 4 5  3 4 4 5  4 5 5 6  3 4 4 5  4 5 5 6  4 5 5 6  5 6 6 7