Create FindAllNumbersDisappearedInArray.java
Question link: https://leetcode.com/problems/find-all-numbers-disappeared-in-an-array/
Approach 1: Mark the indexes as -1, which are encountered. Traverse again, and the ones which are positive are the ones missing.
Approach 2: Traversed linearly, and swapped the value of current index with the value at that particular index, at which this value would have been present. If current index is still not correct,
 then again did the same, and kept doing until it is correct. If the value to be swapped is same, then current index value is set to -2 and moved on. If current index value is correct, move on.