Create BestTimeToBuyAndSellStock.java
Question Link: https://leetcode.com/problems/best-time-to-buy-and-sell-stock/submissions/
Approach: Modified the array to store differences between consecutive elements. Ex: 7 1 5 3 6 4 will become -6 4 -2 4 -2
Then used sliding window technique to find the window with maximum sum. Whenever the window sum<0, window is reset and both start and end point to the index next to the window end.
Because it's useless to carry negative forward, as it will anyways decrease the window sum going forward.
Otherwide only end is moved forward by one.
And after each movement of window, the maximum is compared and updated if required.