Create SortedArrayToBST.java
Question Link: https://leetcode.com/problems/convert-sorted-array-to-binary-search-tree/submissions/
Approach: Solve recursivly. Keep middle element of array as root, and to it's left would be the left side of the array, and to it's right would be the right side of the array.
So, both would be again passed to the recursive function.