Create ValidAnagram.java
Question Link: https://leetcode.com/problems/valid-anagram/
Approach: Use an array of length 26, and traverse the strings and update the index corresponding to the character in the strings. Increment the count for first string and decrement for second.
If all the indices of the array are 0, return true, else return false.