Create JumpGame2.java
Question Link:https://leetcode.com/problems/jump-game-ii/
Approach: For the first index, find the maximum index it can go to. From index 1 till that max index will be at level 2. Then do the same to find level 3, and so on.
The number of levels - 1, is the minimum number of jumps required.
Approach 2: Using DP, but it got TLE in one case of a large sorted array.