Create MinStack.java
Question Link:https://leetcode.com/problems/min-stack/
Approach1: Use a stack. When pushing, check if the minimum needs to be updated. If yes, first push the old minimum and then the number, and update the minimum variable.
When popping, check if the minimum element is being popped out, then pop the one below that and store it in min variable as it is the old minimum.
Approach2: Using a List of Lists. The inner list contains the number, and the minimum value uptil that point.