Create GenerateParentheses.java
Question Link: https://leetcode.com/problems/generate-parentheses/
Approach: Used Recursion. End Case is n=0, in which an empty string is returned. For a length n, the number of combinations will be: "(" + //all strings of length c + ")" + //all strings of length n-c-1, from c = 0, to c = n-1. And adding it to the list.