Create 4SumII.java
Question Link: https://leetcode.com/problems/4sum-ii/
Approach: Found all possible combinations of first 2 arrays, and stored the sum along with their frequency in hashmap. Then found all possible combinations of 3rd and 4th arrays, and checked if the negative of that sum is present
in the hashmap. If yes, the freuqency is added to the result.