Create PathSumIII.java
Question Link: https://leetcode.com/problems/path-sum-iii/
Approach: Create the tree similar to prefix sum array. At each level, pass a hashmap, adding the number, according to it's current value
which is required to be in it's child, so that the required sum is possible.
While coming up, remove that number.