Create InsertDeleteGetRandomO1.java
Question Link: https://leetcode.com/problems/insert-delete-getrandom-o1/
Approach: Create a list, and a HashMap having the mappings of numbers with their index position on the list. On removal, put the last position of the list in the position of element to
be removed, and remove the last element, and update the HashMap accordingly.