Create PallindromePartitioning.java
Question Link: https://leetcode.com/problems/palindrome-partitioning/
Approach: Create a list of list of integers, and for each index, store the next index, between which(both inclusive) the string will be a pallindrome.
Next use recursive approach to find all possible paths to reach the end of the string, while add pallindrominc substrings to a list, and if the end is reached, add that list to the result.