Create SearchInA2DMatrix2.java
Question Link: https://leetcode.com/problems/search-a-2d-matrix-ii/
Approach: Used Recursion. From top left, check on diagonal elements, till the element is lesser than target, then passed the remaining of array on top and right,
and the one on left and below to the same function.
A better approach was found in which it started form top right, and if target is found then return true else if it is lesser, than col--, else row++