Create LongestSubstringWithAtLeastKRepeatingCharacters.java
Question Link: https://leetcode.com/problems/longest-substring-with-at-least-k-repeating-characters/
Approach: The longest subsequence can have characters ranging from 1 from 26. So, for each number of unique characters, slided a window, to find the longest substring with that many
characters, and did it for all unique counts from 1 to 26, and kept updating the max.