Create IncreasingTripletSubsequence.java
Question Link: https://leetcode.com/problems/increasing-triplet-subsequence/
Approach: Kept track of 2 increasing numbers, and if the 3rd number greater than previous is found, return true. Whereas, if another set of 2 numbers which is less than the previous set
is found, then replace the first set with this one, and move on.