Create WordLadder.java
Question Link: https://leetcode.com/problems/word-ladder/
Approach: Change each letter, of current word to every other letter one by one, and if it is present in wordlist it means they are direct neighbours, and add 1 to distance. Keep doing this, until you find the endWord. This is a slight variation of BFS.