Create TrappingRainWater.java
Question Link: https://leetcode.com/problems/trapping-rain-water
Approach: Used 2 pointers from start. Kept 1 pointer fixed, and moved the other to find the next height which is greater or equal to the current height. If found, added the water trapped,
and set the first pointer to the pointer having the next height. If no height is found which is greater or equal, find the max height after the current height. And calculate water
trapped accordingly, then set the first pointer to the index having that height.