Create CountOfSmallerNumbersAfterSelf.java
Question Link: https://leetcode.com/problems/count-of-smaller-numbers-after-self
Approach: Traverse from right to left, and create a BST in that order. So, at each addition of node, count how many numbers you passed were smaller, and also add the leftChildren of those nodes which were greater.
So, we have to keep track of leftChildren as well. Also keep track of duplicate entries and add them to the sum accordingly.