Create MergeKSortedLists.java
Question Link: https://leetcode.com/problems/merge-k-sorted-lists/
Approach: Since all the lists are already sorted, the respective heads of the lists would already be in sorted order. So, we just need to keep track of the heads, and need to know that for now, which is the largest head, instead of knowing the whole sorted order
of all heads at that point of time. And then we need to remove the largest head, and add the next head of that list to that list of heads. For finding the maximum head at a point of time, we use PriorityQueue or to say MaxHeap.