Create LargestRectangleInHistogram.java
Question Link: https://leetcode.com/problems/largest-rectangle-in-histogram/
Approach: For each bar, we need to find the maximum rectangle around it, which is of the same height as itself. So, we first store the index of the first bar to the left and to the right
which is smaller than the current bar in 2 arrays. Then we traverse all the bars, and find the maximum area.