Create MinimumWindowSubstring.java
Question Link: https://leetcode.com/problems/minimum-window-substring/
Approach: Used a sliding window, and kept track of the count of the required letters inside that window. Whenever the count is insufficient, increment end pointer of the window.
Else if the count is sufficient, update the result string, and increment the starting pointer of the window. Keep doing this until the end pointer reaches the end.