Create PartitionLabels.java
Question Link: https://leetcode.com/problems/partition-labels/
Approach: Stored count of all charaters in the string.
Then started traversing by reduscing the count, if count is not zero, add it to a set, else if it is zero remove it from the set.
Wherever the size of set will become zero, we would have a partition.