Create LargestSubarrayOf0sAnd1s.java
Question Link: https://practice.geeksforgeeks.org/problems/largest-subarray-of-0s-and-1s/0/?track=md-arrays&batchId=144#
Approach: Maintained an array which kept numOfZeroesYet-numOfOnesYet. Then in that array, I need those 2 indexes which are farthest
apart, and have the same value. So, I use a hashmap for that. I add the value as key and index as value to hashmap, if it is not there already.
If the value is already present as a key, I find the difference in its current index position and the one in the hashmap and update max.
Also, whenever there is a 0 in the array, the max shall be checked with the lenght of subarray from 0 index till that index.
Finally return max.