Create MinimumPlatforms.java
Question Link: https://practice.geeksforgeeks.org/problems/minimum-platforms-1587115620/0/?track=md-arrays&batchId=144#
Approach: Created an array of length equal to number of minutes in a day, and incremented value at index=minute at which train
enters, and decremented when it leaves. So in this way when we do a prefix sum, we get the number of trains at any instance.