Create SmallestWindowInAStringContainingAllCharactersOfAnotherString.java
Question Link: https://practice.geeksforgeeks.org/problems/smallest-window-in-a-string-containing-all-the-characters-of-another-string-1587115621/0/?track=md-string&batchId=144#
Approach: Maintained count of target character count in an array. Traversed first string with start and end pointer and kept updating the count array and in each traversal
checking if the array satisfies the required condition. If yes, min window length is updated.