Create LongestKUniqueCharactersSubstring.java
Question Link: https://practice.geeksforgeeks.org/problems/longest-k-unique-characters-substring0853/0/?track=md-string&batchId=144
Approach: Used sliding window technique, and a hashmap to store the count of characters in the current window, also a variable to store the current unique count, based on the
counts of hashmap which is updated for a particular character in each iteration. Whenever the count is equal to k, update the max if it is greater then current max.