Create LongestPallindromeInAString.java
Question Link: https://practice.geeksforgeeks.org/problems/longest-palindrome-in-a-string/0/?track=md-string&batchId=144#
Approach: For each character, traverse equally to left and right until the charaters are equal. The maximum such window is the answer. This was for odd length, do the same for even lenght, the only change would be that we have to take 2 characters as centre instead of single.