Create RecursivelyRemoveAllAdjacentDuplicates.java
Question Link: https://practice.geeksforgeeks.org/problems/recursively-remove-all-adjacent-duplicates/0/?track=md-string&batchId=144
Approach: Recursively call functuon with first character and remaining string, check if the char is equal to the first char of remaining string, if no then add it to the string which would be returned in the recursive call.
Also, on finally getting the answer, keep calling the function with updated string until the result is same for 2 consecutive calls. Ex: missippie will become miiie which will become me