Create CoinChange.java
Question Link: https://practice.geeksforgeeks.org/problems/coin-change2448/1#
Approach: Use Dynamic programming. Create a 2D array of rows denoting the number of coins, and columns denoting the amount. For any index, the value in that cell is equal to the number of ways that amount can be reached using or not using that many coins. So, that is equal to the sum of value denoting the amount-that coin value and same row, and the value denoting same amount, but 1 less number of coins. So, dp[i][j] = dp[i][amount - coin[i]] + dp[i-1][amount], where amount at that cell is equal to the column number, j.