Create PrintAnagramsTogether.java
Question Link: https://practice.geeksforgeeks.org/problems/print-anagrams-together/1
Approach: For each string created another string having counts of each alphabet in alphabetical order such that the resultant string is same for anagrams, and stored them as key
in a hashmap with value pointing to a list having indices of array having the string which is the source of that resultant string, hence clubbing all anagrams together. Then just
stored them in an array.