Create MaximumXorOfTwoNumbersInArray.java
Question link: https://practice.geeksforgeeks.org/problems/maximum-xor-of-two-numbers-in-an-array/1
Approach: Used Trie. A trie to store binary numbers. Each node can have two children only. left child mean current bit is 0, and right child means current bit is 1. Root starts with
the most significant bit. So, we insert all elements in the Trie like that. And then for each element, we try to find the max XOR. So the idea is that, starting from the most significant bit,
we start to towards the child which has the opposite bit. Then only XOR for that bit will be 1, and our result will maximize. Suppose current bit is 1, then we will move left, as left
will go towards the numbers which have 0 at that bit, and vice-versa, thus the resultant number will have 1 in that bit position. If the opposite bit path is not there, we have to
move to the same bit path only then, and in that case, the resultant number will have 0 at that bit.