Create PrerequisiteTasks.java
Question Link: https://practice.geeksforgeeks.org/problems/prerequisite-tasks/1/?company[]=Amazon&company[]=Amazon&page=1&query=company[]Amazonpage1company[]Amazon
Approach: Stored dependencies in a hashmap. Key is current task, value is list of dependencies. For each task, we check if dependencies are complete, and if not, we do the same for
each of them. Like this we continue, and if we encounter a cycle, we return false.