Create OccurencesOf2AsADigit.java
Question Link: https://practice.geeksforgeeks.org/problems/occurences-of-2-as-a-digit/1/?company[]=Amazon&company[]=Amazon&difficulty[]=1&page=1&query=company[]Amazondifficulty[]1page1company[]Amazon
Approach: Constraint of N is 100000.
For every 10, number of 2 is 1
For every 100, number of 2 is 20
For every 1000, number of 2 is 300
For every 10000, number of 2 is 4000
For every 100000, number of 2 is 50000
We use this, and have edge cases for scenraios, where we have to count till a number which starts with 2.
Like from 20 to 29, 200 to 299, 2000 to 2999, and so on.