Create NumberOfGoodWaysToSplitAString.java
Question Link: https://leetcode.com/problems/number-of-good-ways-to-split-a-string/
Approach: Store number of unique characters from end to a particular index for both forward and backward pass in 2 arrays. Then in a single traversal, compare number of unique characters on left with that on right for each partition, and increment the result accordingly.