Create DeleteNodesAndReturnForest.java
Question Link: https://leetcode.com/problems/delete-nodes-and-return-forest/
Approach: Store references and parent nodes in hashmaps so that when deleting we can get the references in constant time. Then traverse the list of nodes to delete, and for each of them, add it's left and right child as new trees in the set, and if it itself exists in the set, remove it, and remove it's reference from it's parent.