Create TimeNeededToInformAllEmployees.java
Question Link: https://leetcode.com/problems/time-needed-to-inform-all-employees/
Approach: Add manager to list of subordinate entries in hashmap. Made a recursive function which will take a person as input and return time taken by him and his subordinates to inform. Called the function with the head of the organization. Recursively went to its subordinates and called for each of them. The most time taken amongst them will be the time taken by his subordinates so add his time to it and return.