Create LongestStringChain.java
Question Link: https://leetcode.com/problems/longest-string-chain/
Approach: Stored the strings in list of list based on their lengths. And started from seach string and moved to the next list of strings of length greater by 1, and checked which are successors, and recursively did the same for them. Also stored the answer for each string in a map to use for repeating sub problems.