Create FindDupicateSubtrees.java
Question Link: https://leetcode.com/problems/find-duplicate-subtrees/
Approach: Traversed the tree and stores the references of nodes according to value in HashMap, and traversed again then found all nodes with matching values, and compared them and also stored the comparison result in hashmap dp, for O(1) comparison next time, and if comparison result is true, added it to the result list.
Better Approach: Do a inorder traveral, and keep creating inrder searialized string and store in hashmap as key, with value pointing to the root node for that inorder string. Then finally just get the entries from the hashmap which have more than 1 node.