Create FindAndReplaceInString.java
Question Link: https://leetcode.com/problems/find-and-replace-in-string/
Approach: The main thing here to note is that if we replace a substring, then the indexing of remaining original charaters to the right of that replacement would change, and hence any future operation to the right of the substring won't point to the correct index.
So, what can we do? We can do the operation from right to left. In this way no future operations will be affected by current change. So we first store the source and target based on indices in hashmaps and sort the indices array. Then from right to left we traverse the indices array and make replacement in the original string if necessary.