Create CarFleet.java
Question Link: https://leetcode.com/problems/car-fleet/
Approach: If we need to find the cars that will merge before they reach the end, we need to understand that if a car A is before car B, and takes lesser or equal time to reach target than car B, then they will form a fleet by the time they reach the target.
Now to compare a car with another car having position greater or lesser would be fairly easy if we sort the positions array, and would have stored the speed based on position in a hashmap earlier. Then we just find the time taken by each of them, and from the car closest to target move backwards and compare the current speed with the next car speed and see if they would form a fleet, if they wont, our answer would have one more fleet, and we would continue likewise.
Time complexity: O(NlogN) because of sorting.