Create SnapshotArray.java
Question Link: https://leetcode.com/problems/snapshot-array/
Approach: Create a list of lists. The first list will have the original items, default 0. And the value in list will be an object having the snap id and the value. For each snap we just maintain snap id, so we just increment that variable. While setting a value, we check what was the latest snap in for that index, and if it is outdated, we add a new object with the updated snap. Finally while getting we search the snap id at that index and return the value.