Create StoneGameIII.java
Question Link: https://leetcode.com/problems/stone-game-iii/
Approach: Its is a recursive solution. We implement a function which would take in the index of the array, and return the maximum difference a person starting at that index can get. We do so by comparing the result from 3 possibilities which are choosing 1,2 or 3 stones. If we choose suppose 2 stones, then the difference would be current stone value + next stone value - (recursive function with i+2). We can store the results in a dp array because there are overlapping subproblems.