Create CarFleetII.java
Question Link: https://leetcode.com/problems/car-fleet-ii/
Approach: Start from the right and maintain a min_till_now, if a speed is lesser than min_till_now, then no car on the left of it would ever be able to meet a car on the right of it, and then update min_till_now to current speed which is the new lowest. Then if the current speed is greater than the next and the next is the min_till_now, then just calculate the time and update the res array. Also maintain another array in which we would store the index of the car to which the current car actually merged into. Now if the current car speed is more than the min_till_now, we then find the first car on the right whose speed is lesser than the current car. The current car may directly merge into that, or if it takes more time that that next car to merge into next and change it speed, then it might merge into that next, but then again if that also takes more time, this chain goes on until we find a car which is not going to merge into any next, or the total time taken till now is less that of the the queried car time to merge into its next. Thus we update the time and mergedInto arrays for the current car.