Create SplitArrayLargestSum.java
Question Link: https://leetcode.com/problems/split-array-largest-sum/
Approach: In this type of question in which we need to find something like, find minimum of maximum or find maximum of minimum, then most of the times it is done using a binary search. The max possible partition is sum of all numbers and the min possible is the maximum number in the array, so we would do a binary search in that range to find the minimum such sum for which m partitions are possible.